+++ Converters/Rhino/Speckle.Converters.RhinoShared/ToSpeckle/Meshing/DisplayMeshExtractor.cs
@@ -1,4 +1,5 @@
 using Rhino.DocObjects;
+using Speckle.DoubleNumerics;
 using Speckle.Sdk.Common.Exceptions;
 
 namespace Speckle.Converters.Rhino.ToSpeckle.Meshing;
@@ -95,5 +96,63 @@ public static class DisplayMeshExtractor
     }
 
     return minEdgeLength;
+  }
+
+  /// <summary>
+  /// Extracting Rhino Mesh from Rhino GeometryBase after moving it to origin (if needed).
+  /// </summary>
+  public static RG.Mesh MoveToOriginAndGetDisplayMesh(
+    RG.GeometryBase geometry,
+    bool modelFarFromOrigin,
+    out RG.Vector3d? vectorToOriginalGeometry
+  )
+  {
+    vectorToOriginalGeometry = null;
+
+    // 1. General check: if Model is NOT far from origin (99% of Rhino models): extract meshes as usual
+    if (!modelFarFromOrigin)
+    {
+      return GetGeometryDisplayMesh(geometry, true);
+    }
+    // 2. Geometry check: if the model extent is far from origin, but object itself is NOT far from origin: extract meshes as usual
+    if (!TryGetTranslationVector(geometry, out RG.Vector3d vectorToGeometry))
+    {
+      return GetGeometryDisplayMesh(geometry, true);
+    }
+    // 3. If the object is far from origin and risking faulty meshes due to precision errors: duplicate geometry and move it to origin
+    RG.GeometryBase geometryToMesh = geometry.Duplicate();
+    geometryToMesh.Transform(RG.Transform.Translation(-vectorToGeometry));
+
+    vectorToOriginalGeometry = vectorToGeometry;
+    return GetGeometryDisplayMesh(geometryToMesh, true);
+  }
+
+  /// <summary>
+  /// Getting translation vector from origin to the Geometry bbox Center (if geometry is far from origin and translation needed)
+  /// </summary>
+  /// <returns>
+  /// True and the vector from origin to Geometry bbox center (if translation needed), otherwise false and zero-length vector.
+  /// </returns>
+  private static bool TryGetTranslationVector(RG.GeometryBase geom, out RG.Vector3d vector)
+  {
+    vector = new RG.Vector3d();
+    var geometryBbox = geom.GetBoundingBox(false); // 'false' for 'accurate' parameter to accelerate bbox calculation
+    if (geometryBbox.Min.DistanceTo(RG.Point3d.Origin) > 1e6 || geometryBbox.Max.DistanceTo(RG.Point3d.Origin) > 1e6)
+    {
+      vector = new RG.Vector3d(geometryBbox.Center);
+      return true;
+    }
+
+    return false;
+  }
+
+  public static void MoveSpeckleMeshes(List<SOG.Mesh> displayValue, RG.Vector3d? vectorToGeometry, string units)
+  {
+    if (vectorToGeometry is RG.Vector3d vector)
+    {
+      Matrix4x4 matrix = new(1, 0, 0, vector.X, 0, 1, 0, vector.Y, 0, 0, 1, vector.Z, 0, 0, 0, 1);
+      SO.Transform transform = new() { matrix = matrix, units = units };
+      displayValue.ForEach(x => x.Transform(transform));
+    }
   }
 }
