+++ Converters/Rhino/Speckle.Converters.RhinoShared/ToSpeckle/Raw/HatchToSpeckleConverter.cs
@@ -39,12 +39,18 @@ public class HatchToSpeckleConverter : ITypedConverter<RG.Hatch, SOG.Region>
     // create display mesh from region by converting to brep first
     var brep = RG.Brep.TryConvertBrep(target);
 
-    List<SOG.Mesh> displayValue = ToSpeckleMeshUtility.GetSpeckleMeshes(
+    // Extracting and converting meshes
+    // 1. If needed, move geometry to origin before all operations; extract Rhino Mesh
+    RG.Mesh movedDisplayMesh = DisplayMeshExtractor.MoveToOriginAndGetDisplayMesh(
       brep,
       _settingsStore.Current.ModelFarFromOrigin,
-      _settingsStore.Current.SpeckleUnits,
-      _meshConverter
+      out RG.Vector3d? vectorToGeometry
     );
+    // 2. Convert extracted Mesh to Speckle. We don't move geometry back yet, because 'far from origin' geometry is causing Speckle conversion issues too
+    List<SOG.Mesh> displayValue = new() { _meshConverter.Convert(movedDisplayMesh) };
+    // 3. Move Speckle geometry back from origin, if translation was applied
+    DisplayMeshExtractor.MoveSpeckleMeshes(displayValue, vectorToGeometry, _settingsStore.Current.SpeckleUnits);
+
     return new SOG.Region
     {
       boundary = boundary,
